remove English strings from language names
Having this string is un-necessary in the case of the default English UI
(if you don't recognize the name of the language, content in that
language is not going to be very useful to you) and outright dead-weight
in the context of a localized UI (if you are using a not-English
version, non-localized strings hard coded in English are not that
useful).

Unless we include a full language name matrix with each language name
localized in each of the languages the site can be localized, it is more
generally useful to just show languages using their localized names (not to
mention easier).

Looking around at major platforms with localization support, this
appears to be the norm now (as opposed to a few years ago when
localization was first rolling out to a few programs and was being
driven by and English-first mindset).

If anything needs a hard-code English fall-back for people to "escape"
from a language they do not understand it would be the name of the
setting and the widget itself. Once they find the widget, having
"English" as a value in the list is enough.